[cleanup] Abstract setup_city_gates into gates.py
